refactor: move VN p2p rpc definition into common crate
