+++ applications/tari_indexer/Cargo.toml
@@ -15,14 +15,13 @@ tari_common = { git = "https://github.com/tari-project/tari.git", tag = "v0.50.0
 tari_common_types = { git = "https://github.com/tari-project/tari.git", tag = "v0.50.0-pre.0", package = "tari_common_types" }
 tari_comms = { git = "https://github.com/tari-project/tari.git", tag = "v0.50.0-pre.0", package = "tari_comms" }
 tari_comms_logging = { path = "../../comms/tari_comms_logging" }
-tari_comms_rpc_macros = { git = "https://github.com/tari-project/tari.git", tag = "v0.50.0-pre.0", package = "tari_comms_rpc_macros" }
 tari_core = { git = "https://github.com/tari-project/tari.git", tag = "v0.50.0-pre.0", package = "tari_core", default-features = false, features = ["transactions"] }
 tari_crypto = { version="0.16.10" }
 tari_p2p = { git = "https://github.com/tari-project/tari.git", tag = "v0.50.0-pre.0", package = "tari_p2p" }
 tari_shutdown = { git = "https://github.com/tari-project/tari.git", tag = "v0.50.0-pre.0", package = "tari_shutdown" }
 tari_storage = { git = "https://github.com/tari-project/tari.git", tag = "v0.50.0-pre.0", package = "tari_storage" }
 
-tari_dan_app_grpc = { path = "../tari_dan_app_grpc" }
+tari_validator_node_rpc = { path = "../../dan_layer/validator_node_rpc" }
 tari_dan_app_utilities = { path = "../tari_dan_app_utilities" }
 tari_dan_common_types = { path = "../../dan_layer/common_types" }
 tari_dan_core = { path = "../../dan_layer/core" }
@@ -32,7 +31,6 @@ tari_dan_storage_sqlite = { path = "../../dan_layer/storage_sqlite" }
 tari_engine_types = { path = "../../dan_layer/engine_types" }
 tari_template_lib = { path = "../../dan_layer/template_lib" }
 tari_transaction = { path = "../../dan_layer/transaction" }
-tari_validator_node = { path = "../tari_validator_node" }
 tari_validator_node_client = { path = "../../clients/validator_node_client" }
 
 anyhow = "1.0.53"
