+++ applications/tari_indexer/src/dan_layer_scanner.rs
@@ -24,12 +24,12 @@ use futures::StreamExt;
 use log::*;
 use serde::{Deserialize, Serialize};
 use tari_crypto::ristretto::RistrettoPublicKey;
-use tari_dan_app_grpc::proto::rpc::VnStateSyncResponse;
 use tari_dan_app_utilities::epoch_manager::EpochManagerHandle;
 use tari_dan_common_types::{Epoch, ShardId};
 use tari_dan_core::services::{epoch_manager::EpochManager, ValidatorNodeClientFactory};
 use tari_engine_types::substate::{Substate, SubstateAddress};
 use tari_template_lib::{models::NonFungibleIndexAddress, prelude::ResourceAddress};
+use tari_validator_node_rpc::proto::rpc::VnStateSyncResponse;
 
 use crate::p2p::services::rpc_client::TariCommsValidatorNodeClientFactory;
 
@@ -227,8 +227,8 @@ impl DanLayerScanner {
         let mut sync_vn_rpc_client = sync_vn_client.create_connection().await?;
 
         // request the shard substate to the VN
-        let shard_id_proto: tari_dan_app_grpc::proto::common::ShardId = shard_id.into();
-        let request = tari_dan_app_grpc::proto::rpc::VnStateSyncRequest {
+        let shard_id_proto: tari_validator_node_rpc::proto::common::ShardId = shard_id.into();
+        let request = tari_validator_node_rpc::proto::rpc::VnStateSyncRequest {
             start_shard_id: Some(shard_id_proto.clone()),
             end_shard_id: Some(shard_id_proto),
             inventory: vec![],
