+++ applications/tari_indexer/src/p2p/mod.rs
@@ -20,7 +20,4 @@
 // WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-mod rpc;
-pub use rpc::create_validator_node_rpc_service;
-
 pub mod services;
