--- applications/tari_indexer/src/p2p/rpc/service_impl.rs
@@ -1,100 +1,0 @@
-//  Copyright 2023, The Tari Project
-//
-//  Redistribution and use in source and binary forms, with or without modification, are permitted provided that
-// the  following conditions are met:
-//
-//  1. Redistributions of source code must retain the above copyright notice, this list of conditions and the
-// following  disclaimer.
-//
-//  2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the
-//  following disclaimer in the documentation and/or other materials provided with the distribution.
-//
-//  3. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote
-//  products derived from this software without specific prior written permission.
-//
-//  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED
-// WARRANTIES,  INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
-// PARTICULAR PURPOSE ARE  DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY
-// DIRECT, INDIRECT, INCIDENTAL,  SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
-// PROCUREMENT OF SUBSTITUTE GOODS OR  SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-// CAUSED AND ON ANY THEORY OF LIABILITY,  WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
-// OTHERWISE) ARISING IN ANY WAY OUT OF THE  USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH
-// DAMAGE.
-
-use log::*;
-use tari_comms::protocol::rpc::{Request, Response, RpcStatus, Streaming};
-use tari_dan_app_grpc::{
-    proto,
-    proto::rpc::{VnStateSyncRequest, VnStateSyncResponse},
-};
-use tari_dan_common_types::NodeAddressable;
-use tari_dan_core::services::PeerProvider;
-use tokio::{sync::mpsc, task};
-
-const LOG_TARGET: &str = "tari::dan::p2p::rpc";
-
-use crate::p2p::rpc::ValidatorNodeRpcService;
-
-pub struct ValidatorNodeRpcServiceImpl<TPeerProvider> {
-    peer_provider: TPeerProvider,
-}
-
-impl<TPeerProvider: PeerProvider> ValidatorNodeRpcServiceImpl<TPeerProvider> {
-    pub fn new(peer_provider: TPeerProvider) -> Self {
-        Self { peer_provider }
-    }
-}
-
-#[tari_comms::async_trait]
-impl<TPeerProvider> ValidatorNodeRpcService for ValidatorNodeRpcServiceImpl<TPeerProvider>
-where TPeerProvider: PeerProvider + Clone + Send + Sync + 'static
-{
-    async fn submit_transaction(
-        &self,
-        _request: Request<proto::rpc::SubmitTransactionRequest>,
-    ) -> Result<Response<proto::rpc::SubmitTransactionResponse>, RpcStatus> {
-        todo!()
-    }
-
-    async fn get_peers(
-        &self,
-        _request: Request<proto::rpc::GetPeersRequest>,
-    ) -> Result<Streaming<proto::rpc::GetPeersResponse>, RpcStatus> {
-        let (tx, rx) = mpsc::channel(100);
-        let peer_provider = self.peer_provider.clone();
-
-        task::spawn(async move {
-            let mut peer_iter = peer_provider.peers_for_current_epoch_iter().await;
-            while let Some(Ok(peer)) = peer_iter.next() {
-                // if peer.identity_signature.is_none() {
-                //     continue;
-                // }
-                if tx
-                    .send(Ok(proto::rpc::GetPeersResponse {
-                        identity: peer.identity.as_bytes().to_vec(),
-                        // identity_signature: peer.identity_signature.map(Into::into),
-                        addresses: peer.addresses.iter().map(|(a, _)| a.to_vec()).collect(),
-                        claims: peer.addresses.into_iter().map(|(_, c)| c.into()).collect(),
-                    }))
-                    .await
-                    .is_err()
-                {
-                    debug!(
-                        target: LOG_TARGET,
-                        "Peer stream closed by client before completing. Aborting"
-                    );
-                    break;
-                }
-            }
-        });
-
-        Ok(Streaming::new(rx))
-    }
-
-    async fn vn_state_sync(
-        &self,
-        _request: Request<VnStateSyncRequest>,
-    ) -> Result<Streaming<VnStateSyncResponse>, RpcStatus> {
-        todo!()
-    }
-}
