+++ applications/tari_indexer/src/p2p/services/mod.rs
@@ -20,7 +20,6 @@
 // WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-pub mod comms_peer_provider;
 pub mod epoch_manager;
 pub mod rpc_client;
 pub mod template_manager;
