+++ applications/tari_validator_node/Cargo.toml
@@ -18,7 +18,7 @@ tari_comms_logging = { path = "../../comms/tari_comms_logging" }
 tari_comms_rpc_macros = { git = "https://github.com/tari-project/tari.git", tag = "v0.50.0-pre.0" }
 tari_core = { git = "https://github.com/tari-project/tari.git", tag = "v0.50.0-pre.0", default-features = false, features = ["transactions"] }
 tari_crypto ={ version = "0.16.10" }
-tari_dan_app_grpc = { path = "../tari_dan_app_grpc" }
+tari_validator_node_rpc = { path = "../../dan_layer/validator_node_rpc" }
 tari_dan_app_utilities = { path = "../tari_dan_app_utilities" }
 tari_dan_common_types = { path = "../../dan_layer/common_types" }
 tari_dan_core = { path = "../../dan_layer/core" }
