+++ applications/tari_validator_node/src/comms/deserialize.rs
@@ -31,8 +31,8 @@ use prost::Message;
 use tari_comms::{message::InboundMessage, types::CommsPublicKey, PeerManager};
 use tari_comms_logging::SqliteMessageLog;
 use tari_crypto::tari_utilities::ByteArray;
-use tari_dan_app_grpc::proto;
 use tari_dan_core::{message::DanMessage, models::TariDanPayload};
+use tari_validator_node_rpc::proto;
 use tower::{Service, ServiceExt};
 const LOG_TARGET: &str = "tari::validator_node::comms::messaging";
