+++ applications/tari_validator_node/src/dry_run_transaction_processor.rs
@@ -29,7 +29,6 @@ use std::{
 use futures::StreamExt;
 use log::info;
 use tari_comms::{protocol::rpc::RpcStatus, NodeIdentity};
-use tari_dan_app_grpc::proto::rpc::VnStateSyncResponse;
 use tari_dan_app_utilities::epoch_manager::EpochManagerHandle;
 use tari_dan_common_types::{Epoch, ObjectPledge, PayloadId, ShardId, SubstateState};
 use tari_dan_core::{
@@ -53,6 +52,7 @@ use tari_engine_types::{
     substate::{Substate, SubstateAddress},
 };
 use tari_transaction::{SubstateChange, Transaction};
+use tari_validator_node_rpc::proto::rpc::VnStateSyncResponse;
 use thiserror::Error;
 
 use crate::{
@@ -255,8 +255,8 @@ impl DryRunTransactionProcessor {
             };
 
             // request the shard substate to the VN
-            let shard_id_proto: tari_dan_app_grpc::proto::common::ShardId = shard_id.into();
-            let request = tari_dan_app_grpc::proto::rpc::VnStateSyncRequest {
+            let shard_id_proto: tari_validator_node_rpc::proto::common::ShardId = shard_id.into();
+            let request = tari_validator_node_rpc::proto::rpc::VnStateSyncRequest {
                 start_shard_id: Some(shard_id_proto.clone()),
                 end_shard_id: Some(shard_id_proto),
                 inventory: vec![],
