+++ applications/tari_validator_node/src/p2p/rpc/mod.rs
@@ -23,36 +23,13 @@
 mod service_impl;
 
 pub use service_impl::ValidatorNodeRpcServiceImpl;
-use tari_comms::protocol::rpc::{Request, Response, RpcStatus, Streaming};
-use tari_comms_rpc_macros::tari_rpc;
-use tari_dan_app_grpc::proto;
 use tari_dan_core::services::PeerProvider;
 use tari_dan_storage_sqlite::sqlite_shard_store_factory::SqliteShardStore;
+use tari_validator_node_rpc::rpc_service::ValidatorNodeRpcServer;
 
 use crate::p2p::services::mempool::MempoolHandle;
 
-#[tari_rpc(protocol_name = b"t/vn/1", server_struct = ValidatorNodeRpcServer, client_struct = ValidatorNodeRpcClient)]
-pub trait ValidatorNodeRpcService: Send + Sync + 'static {
-    #[rpc(method = 1)]
-    async fn submit_transaction(
-        &self,
-        request: Request<proto::rpc::SubmitTransactionRequest>,
-    ) -> Result<Response<proto::rpc::SubmitTransactionResponse>, RpcStatus>;
-
-    #[rpc(method = 2)]
-    async fn get_peers(
-        &self,
-        request: Request<proto::rpc::GetPeersRequest>,
-    ) -> Result<Streaming<proto::rpc::GetPeersResponse>, RpcStatus>;
-
-    #[rpc(method = 3)]
-    async fn vn_state_sync(
-        &self,
-        request: Request<proto::rpc::VnStateSyncRequest>,
-    ) -> Result<Streaming<proto::rpc::VnStateSyncResponse>, RpcStatus>;
-}
-
-pub fn create_validator_node_rpc_service<TPeerProvider>(
+pub fn create_tari_validator_node_rpc_service<TPeerProvider>(
     peer_provider: TPeerProvider,
     shard_store_store: SqliteShardStore,
     mempool: MempoolHandle,
