+++ applications/tari_validator_node/src/p2p/rpc/service_impl.rs
@@ -24,10 +24,6 @@ use std::convert::{TryFrom, TryInto};
 
 use log::*;
 use tari_comms::protocol::rpc::{Request, Response, RpcStatus, Streaming};
-use tari_dan_app_grpc::{
-    proto,
-    proto::rpc::{VnStateSyncRequest, VnStateSyncResponse},
-};
 use tari_dan_common_types::{NodeAddressable, ShardId};
 use tari_dan_core::{
     services::PeerProvider,
@@ -35,11 +31,16 @@ use tari_dan_core::{
 };
 use tari_dan_storage_sqlite::sqlite_shard_store_factory::SqliteShardStore;
 use tari_transaction::Transaction;
+use tari_validator_node_rpc::{
+    proto,
+    proto::rpc::{VnStateSyncRequest, VnStateSyncResponse},
+    rpc_service::ValidatorNodeRpcService,
+};
 use tokio::{sync::mpsc, task};
 
 const LOG_TARGET: &str = "tari::dan::p2p::rpc";
 
-use crate::p2p::{rpc::ValidatorNodeRpcService, services::mempool::MempoolHandle};
+use crate::p2p::services::mempool::MempoolHandle;
 
 pub struct ValidatorNodeRpcServiceImpl<TPeerProvider> {
     peer_provider: TPeerProvider,
