+++ dan_layer/template_lib/Cargo.toml
@@ -14,7 +14,6 @@ ciborium = { git = "https://github.com/enarx/ciborium.git", rev = "ac38b3dddeadf
 newtype-ops = "0.1.4"
 serde = { version = "1.0.143", default-features = false, features = ["derive", "alloc"] }
 serde-big-array = "0.5.1"
-hex = { version = "0.4.3", default-features = false, features = ["serde"] }
 
 [dev-dependencies]
 serde_json = "1.0.73"
@@ -23,4 +22,4 @@ serde_json = "1.0.73"
 default = ["macro"]
 macro = ["tari_template_macros"]
 # TODO: Proper support for no_std
-std = ["hex/std", "serde/std"]
+std = ["serde/std"]
