+++ dan_layer/validator_node_rpc/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "tari_dan_app_grpc"
+name = "tari_validator_node_rpc"
 authors = ["The Tari Development Community"]
 description = "This crate is to provide a single source for all cross application grpc files and conversions to and from tari-dan applications"
 repository = "https://github.com/tari-project/tari"
@@ -8,8 +8,9 @@ version = "0.50.0-pre.0"
 edition = "2018"
 
 [dependencies]
-tari_common_types = { git = "https://github.com/tari-project/tari.git", tag = "v0.50.0-pre.0", package = "tari_common_types" }
-tari_comms = { git = "https://github.com/tari-project/tari.git", tag = "v0.50.0-pre.0", package = "tari_comms" }
+tari_common_types = { git = "https://github.com/tari-project/tari.git", tag = "v0.50.0-pre.0" }
+tari_comms = { git = "https://github.com/tari-project/tari.git", tag = "v0.50.0-pre.0" }
+tari_comms_rpc_macros = { git = "https://github.com/tari-project/tari.git", tag = "v0.50.0-pre.0" }
 tari_crypto  =  { version = "0.16.10" }
 
 tari_dan_common_types = { path = "../../dan_layer/common_types" }
