+++ dan_layer/validator_node_rpc/proto/common.proto
