+++ dan_layer/validator_node_rpc/src/lib.rs
@@ -22,4 +22,5 @@
 
 pub mod conversions;
 pub mod proto;
+pub mod rpc_service;
 mod utils;
