+++ dan_layer/validator_node_rpc/src/utils.rs
