+++ dan_layer/wallet/storage_sqlite/Cargo.toml
@@ -16,7 +16,6 @@ tari_template_lib = { path = "../../template_lib" }
 tari_transaction = { path = "../../transaction" }
 tari_utilities = "0.4.10"
 
-anyhow = "1.0"
 serde = "1.0.126"
 serde_json = "1.0.92"
 diesel = { version = "2", features = ["sqlite", "chrono"] }
@@ -30,3 +29,9 @@ libsqlite3-sys = { version = "0.25", features = ["bundled"] }
 
 [dev-dependencies]
 tari_dan_common_types = { path = "../../common_types" }
+
+[package.metadata.cargo-machete]
+ignored = [
+    # We want to bundle this lib
+    "libsqlite3-sys",
+]
